pyproject.toml
[project]
name = "fastomop"
version = "0.1.0"
description = "OMOP. Fast..."
readme = "README.md"
license = "Apache-2.0"
keywords = [
    "llm",
    "mcp server",
    "ohdsi",
    "omop",
    "a2a",
    "a2a server",
    "fastomop",
    "evaluation",
    "monitoring",
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]


authors = [{ name = "vvcb" }]
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[sql]>=0.2.14",
    "arize-phoenix-otel>=0.12.1",
    "fasta2a>=0.5.0",
    "fastmcp>=2.10.6",
    "httpx>=0.28.1",
    "ibis-framework[duckdb]>=10.6.0",
    "mcp[cli]>=1.12.0",
    "openinference-instrumentation-mcp>=1.3.0",
    "openinference-instrumentation-openai>=0.1.30",
    "openinference-instrumentation-pydantic-ai>=0.1.3",
    "openinference-semantic-conventions>=0.1.21",
    "opentelemetry-api>=1.35.0",
    "opentelemetry-exporter-otlp>=1.35.0",
    "opentelemetry-sdk>=1.35.0",
    "pydantic-ai>=0.4.4",
    "pydantic-settings>=2.10.1",
]

[project.scripts]
fastomop = "fastomop:main.main"
fastomop_mcp_sql = "fastomop.mcp.sql.server:main"

[project.optional-dependencies]


mcp-postgres = ["ibis-framework[postgres]>=10.6.0", "psycopg>=3.2.9"]

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.15",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "ruff>=0.12.4",
]
